refactor: make Right & Left package-private
