+++ src/main/java/org/fungover/breeze/funclib/control/Either.java
@@ -1,6 +1,7 @@
 package org.fungover.breeze.funclib.control;
 
 import java.io.Serializable;
+import java.util.Objects;
 import java.util.function.Function;
 
 /**
@@ -25,7 +26,7 @@ import java.util.function.Function;
  * @param <L> The type of the left value (typically an error type)
  * @param <R> The type of the right value (typically a success type)
  */
-public abstract sealed class Either<L, R> implements Serializable permits Left, Right{
+public abstract sealed class Either<L extends Serializable, R extends Serializable> implements Serializable permits Left, Right{
 
   /**
    * Checks if this instance represents a left value.
@@ -65,7 +66,7 @@ public abstract sealed class Either<L, R> implements Serializable permits Left, 
    * @param <U>    The new type of the right value.
    * @return A new {@code Either} instance with the transformed right value.
    */
-  public abstract <U> Either<L, U> map(Function<R, U> mapper);
+  public abstract <U extends Serializable> Either<L, U> map(Function<R, U> mapper);
 
   /**
    * Transforms the left value using the provided function if this is a {@code Left},
@@ -75,7 +76,7 @@ public abstract sealed class Either<L, R> implements Serializable permits Left, 
    * @param <U>    The new type of the left value.
    * @return A new {@code Either} instance with the transformed left value.
    */
-  public abstract <U> Either<U, R> mapLeft(Function<L, U> mapper);
+  public abstract <U extends Serializable> Either<U, R> mapLeft(Function<L, U> mapper);
 
   /**
    * Swaps a {@code Right} to a {@code Left} and vice versa.
@@ -92,7 +93,7 @@ public abstract sealed class Either<L, R> implements Serializable permits Left, 
    *
    * @return A new {@code Either} instance with the transformed right value
    */
-  public abstract <U> Either<L, U> flatMap(Function<R, Either<L, U>> mapper);
+  public abstract <U extends Serializable> Either<L, U> flatMap(Function<R, Either<L, U>> mapper);
 
   /** Transforms both the left and right values using the provided functions
    *
@@ -100,7 +101,7 @@ public abstract sealed class Either<L, R> implements Serializable permits Left, 
    * @param rightMapper The function to apply to the right value
    * @param <U> The new type of the left and right values
    *
-   * @return A new {@code Either} instance with the transformed value
+   * @return The result of applying the corresponding mapper function
    */
   public abstract <U> U fold(Function<L, U> leftMapper, Function<R, U> rightMapper);
 
@@ -112,7 +113,7 @@ public abstract sealed class Either<L, R> implements Serializable permits Left, 
    * @param <R>   The type of the right value.
    * @return An {@code Either} instance containing the left value.
    */
-  public static <L, R> Either<L, R> left(L value) {
+  public static <L extends Serializable, R extends Serializable> Either<L, R> left(L value) {
     return new Left<>(value);
   }
 
@@ -124,7 +125,341 @@ public abstract sealed class Either<L, R> implements Serializable permits Left, 
    * @param <R>   The type of the right value.
    * @return An {@code Either} instance containing the right value.
    */
-  public static <L, R> Either<L, R> right(R value) {
+  public static <L extends Serializable, R extends Serializable> Either<L, R> right(R value) {
     return new Right<>(value);
+  }
+}
+
+/**
+ * A concrete implementation of {@code Either} representing a left value.
+ *
+ * <p>By convention, a {@code Left} value represents an error or failure case
+ * in computations that may return one of two possible results.</p>
+ *
+ * <h2>Usage Example:</h2>
+ * <pre>
+ * Either<String, Integer> result = Either.left("Error: Invalid input");
+ * if (result.isLeft()) {
+ *     System.out.println("Failure: " + result.getLeft());
+ * }
+ * </pre>
+ *
+ * <h2>Thread Safety:</h2>
+ * <p>Instances of {@code Left} are immutable and therefore thread-safe.</p>
+ *
+ * @param <L> The type of the left value (typically an error type)
+ * @param <R> The type of the right value (unused in this case)
+ */
+final class Left<L extends Serializable, R extends Serializable> extends Either<L, R> {
+  private final L value;
+
+  /**
+   * Constructs a new {@code Left} instance with the specified value.
+   *
+   * @param value The left value, representing a failure or error.
+   */
+  public Left(L value) {
+    this.value = value;
+  }
+
+  /**
+   * Checks if this instance represents a left value.
+   *
+   * @return {@code true}, since this is always a left value.
+   */
+  @Override
+  public boolean isLeft() {
+    return true;
+  }
+
+  /**
+   * Checks if this instance represents a right value.
+   *
+   * @return {@code false}, since this is always a left value.
+   */
+  @Override
+  public boolean isRight() {
+    return false;
+  }
+
+  /**
+   * Retrieves the left value.
+   *
+   * @return The left value.
+   */
+  @Override
+  public L getLeft() {
+    return value;
+  }
+
+  /**
+   * Throws an exception because a {@code Left} does not contain a right value.
+   *
+   * @return Never returns normally.
+   * @throws UnsupportedOperationException Always thrown when called.
+   */
+  @Override
+  public R getRight() {
+    throw new UnsupportedOperationException("Can not get right value on Left");
+  }
+
+  /**
+   * Returns this instance unchanged, as mapping applies only to right values.
+   *
+   * @param mapper The function to apply to the right value (not used).
+   * @param <U>    The new type of the right value.
+   * @return This {@code Left} instance unchanged.
+   */
+  @Override
+  public <U extends Serializable> Either<L, U> map(Function<R, U> mapper) {
+    return Either.left(value);
+  }
+
+  /**
+   * Transforms the left value using the provided function.
+   *
+   * @param mapper The function to apply to the left value.
+   * @param <U>    The new type of the left value.
+   * @return A new {@code Left} instance with the transformed left value.
+   */
+  @Override
+  public <U extends Serializable> Either<U, R> mapLeft(Function<L, U> mapper) {
+    return Either.left(mapper.apply(value));
+  }
+
+  /**
+   * Swaps the value from {@code Left} to {@code Right}
+   *
+   * @return A new {@code Right} instance with the Left's value
+   */
+  @Override
+  public Either<R, L> swap() {
+    return Either.right(value);
+  }
+
+  /**
+   *
+   * @param mapper The function to apply to the right value
+   * @param <U> The new type of the right value
+   *
+   * @return A new {@code Left} instance with unchanged value
+   */
+  @Override
+  public <U extends Serializable> Either<L, U> flatMap(Function<R, Either<L, U>> mapper) {
+    return Either.left(value);
+  }
+
+  /**
+   *
+   * @param leftMapper The function to apply to the left value
+   * @param rightMapper The function to apply to the right value
+   * @param <U> The new type of the left value
+   *
+   * @return The result of applying the corresponding mapper function
+   */
+  @Override
+  public <U> U fold(Function<L, U> leftMapper, Function<R, U> rightMapper) {
+    return leftMapper.apply(value);
+  }
+
+  /**
+   * Compares this instance with another for equality.
+   *
+   * @param o The object to compare with.
+   * @return {@code true} if the other object is also a {@code Left} with the same value, otherwise {@code false}.
+   */
+  @Override
+  public boolean equals(Object o) {
+    if (!(o instanceof Left<?, ?> left)) return false;
+    return Objects.equals(value, left.value);
+  }
+
+  /**
+   * Computes the hash code for this instance.
+   *
+   * @return The hash code of the left value.
+   */
+  @Override
+  public int hashCode() {
+    return Objects.hashCode(value);
+  }
+
+  /**
+   * Returns a string representation of this instance.
+   *
+   * @return A string describing this {@code Left} instance.
+   */
+  @Override
+  public String toString() {
+    return "Left { " +
+            "value = " + value +
+            " }";
+  }
+}
+
+/**
+ * A concrete implementation of {@code Either} representing a right value.
+ *
+ * <p>By convention, a {@code Right} value represents a success case
+ * in computations that may return one of two possible results.</p>
+ *
+ * <h2>Usage Example:</h2>
+ * <pre>
+ * Either<String, Integer> result = Either.right(42);
+ * if (result.isRight()) {
+ *     System.out.println("Success: " + result.getRight());
+ * }
+ * </pre>
+ *
+ * <h2>Thread Safety:</h2>
+ * <p>Instances of {@code Right} are immutable and therefore thread-safe.</p>
+ *
+ * @param <L> The type of the left value (unused in this case)
+ * @param <R> The type of the right value (typically a success type)
+ */
+final class Right<L extends Serializable, R extends Serializable> extends Either<L, R> {
+  private final R value;
+
+  /**
+   * Constructs a new {@code Right} instance with the specified value.
+   *
+   * @param value The right value, representing a success.
+   */
+  public Right(R value) {
+    this.value = value;
+  }
+
+  /**
+   * Checks if this instance represents a left value.
+   *
+   * @return {@code false}, since this is always a right value.
+   */
+  @Override
+  public boolean isLeft() {
+    return false;
+  }
+
+  /**
+   * Checks if this instance represents a right value.
+   *
+   * @return {@code true}, since this is always a right value.
+   */
+  @Override
+  public boolean isRight() {
+    return true;
+  }
+
+  /**
+   * Throws an exception because a {@code Right} does not contain a left value.
+   *
+   * @return Never returns normally.
+   * @throws UnsupportedOperationException Always thrown when called.
+   */
+  @Override
+  public L getLeft() {
+    throw new UnsupportedOperationException("Can not get left value on Right");
+  }
+
+  /**
+   * Retrieves the right value.
+   *
+   * @return The right value.
+   */
+  @Override
+  public R getRight() {
+    return value;
+  }
+
+  /**
+   * Transforms the right value using the provided function.
+   *
+   * @param mapper The function to apply to the right value.
+   * @param <U>    The new type of the right value.
+   * @return A new {@code Right} instance with the transformed right value.
+   */
+  @Override
+  public <U extends Serializable> Either<L, U> map(Function<R, U> mapper) {
+    return Either.right(mapper.apply(value));
+  }
+
+  /**
+   * Returns this instance unchanged, as mapLeft applies only to left values.
+   *
+   * @param mapper The function to apply to the left value (not used).
+   * @param <U>    The new type of the left value.
+   * @return This {@code Right} instance unchanged.
+   */
+  @Override
+  public <U extends Serializable> Either<U, R> mapLeft(Function<L, U> mapper) {
+    return Either.right(value);
+  }
+
+  /**
+   * Swaps the value from {@code Right} to {@code Left}
+   *
+   * @return A new {@code Left} instance with the Right's value
+   */
+  @Override
+  public Either<R, L> swap() {
+    return Either.left(value);
+  }
+
+  /**
+   *
+   * @param mapper The function to apply to the right value
+   * @param <U> The new type of the right value
+   *
+   * @return A new {@code Right} instance with the transformed right value.
+   */
+  @Override
+  public <U extends Serializable> Either<L, U> flatMap(Function<R, Either<L, U>> mapper) {
+    return mapper.apply(value);
+  }
+
+  /**
+   *
+   * @param leftMapper The function to apply to the left value
+   * @param rightMapper The function to apply to the right value
+   * @param <U> The new type of the right value
+   *
+   * @return The result of applying the corresponding mapper function
+   */
+  @Override
+  public <U> U fold(Function<L, U> leftMapper, Function<R, U> rightMapper) {
+    return rightMapper.apply(value);
+  }
+
+  /**
+   * Compares this instance with another for equality.
+   *
+   * @param o The object to compare with.
+   * @return {@code true} if the other object is also a {@code Right} with the same value, otherwise {@code false}.
+   */
+  @Override
+  public boolean equals(Object o) {
+    if (!(o instanceof Right<?, ?> right)) return false;
+    return Objects.equals(value, right.value);
+  }
+
+  /**
+   * Computes the hash code for this instance.
+   *
+   * @return The hash code of the right value.
+   */
+  @Override
+  public int hashCode() {
+    return Objects.hashCode(value);
+  }
+
+  /**
+   * Returns a string representation of this instance.
+   *
+   * @return A string describing this {@code Right} instance.
+   */
+  @Override
+  public String toString() {
+    return "Right { " +
+            "value = " + value +
+            " }";
   }
 }
